+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-- fix all material items converting to ingots when they cool
+- re-enable remapper for `gregitas` namespace ore blocks
+- fix EBF heat number showing the wrong value
+- fix EBF heat number not being formatted correctly